+++ src/pages/toolDetail/components/toolIntro/ToolIntro.stories.tsx
@@ -1,5 +1,5 @@
+import { TOOL_DESCRIPTIONS } from '@pages/toolDetail/mocks/description';
 import type { Meta, StoryFn } from '@storybook/react';
-import { TOOL_DESCRIPTIONS } from 'src/pages/toolDetail/constants/description';
 
 import ToolIntro from './ToolIntro';
