+++ src/pages/toolDetail/mocks/coreFeature.ts
@@ -1,0 +1,33 @@
+export const TOOL_COREFEATURES = {
+  coreId53: {
+    coreTitle: '서버/채널 관리',
+    coreContent:
+      '사용자는 자체 서버를 생성하거나 다른 사용자가 만든 서버에 참여할 수 있으며, 서버 내에서 여러 채널을 문자 및 음성 형식으로 생성하여 각각의 채널을 특정 주제나 목적에 따라 구분할 수 있습니다.',
+  },
+  coreId54: {
+    coreId: 54,
+    coreTitle: '음성/영상 통화',
+    coreContent:
+      '디스코드는 음성 채팅과 영상 통화를 지원하여 친구들과 실시간 대화를 나누며 게임을 하거나, 온라인 회의를 할 때 유용합니다. ',
+  },
+  coreId55: {
+    coreId: 55,
+    coreTitle: '파일 전송/화면 공유',
+    coreContent:
+      '디스코드는 파일 전송과 화면 공유 기능도 제공하여, 문서나 이미지를 공유하거나 자신의 화면을 실시간으로 보여줄 수 있습니다. ',
+  },
+  coreId56: {
+    coreTitle: '정보 검색 및 제공',
+    coreContent: '다루다는 다르다. 정보검색 불가능 but 정보제공 쌉가능!',
+  },
+  coreId57: {
+    coreId: 54,
+    coreTitle: '음성 인식 및 반응',
+    coreContent: '다루다는 다르다. 음성인식 불가능 but 빠른 반응 쌉가능!',
+  },
+  coreId58: {
+    coreId: 55,
+    coreTitle: '이미지 생성 및 분석',
+    coreContent: '다루다는 다르다. 이미지 생성 불가능 but 잘나온 사진 분석 쌉가능!',
+  },
+};
